8144019: added TieredCompilation
https://bugs.openjdk.java.net/browse/JDK-8144019

+ a PPC port of the following community change to C1:
  http://hg.openjdk.java.net/jdk8u/jdk8u/hotspot/rev/0b85ccd62409

+++ hotspot/src/share/vm/c1/c1_Compilation.hpp
@@ -194,7 +194,7 @@ class Compilation: public StackObj {
     return (int) NMethodSizeLimit;  // default 256K or 512K
 #else
     // conditional branches on PPC are restricted to 16 bit signed
-    return MIN2((unsigned int)NMethodSizeLimit,32*K);
+    return MIN2((unsigned int)NMethodSizeLimit,(unsigned int)(32*K));
 #endif
   }
   static int desired_max_constant_size() {
